Component locking feature
Added when you move on object to a desired location you can "lock" it to the object similar to the scanner from the horror game and i just remembered that game and fuck.

SCRIPTS ADDED;
ComponentLocking - detects when an object with tag in inside and after clicking the surrounding area of the pad locks and unlocks the object

SCRIPTS ALTERED;
Interaction - added code at the bottom that accesses the script above to lock the object in place.